requirements.txt
streamlit==1.38.0
Pillow==10.2.0
torch==2.2.1
requests>=2.31.0
transformers>=4.45.2  
safetensors>=0.4.5  

